Move client selector into its own component and rearrange sidebar some more
