+++ client/components/Navbar.jsx
@@ -1,34 +1,12 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import {
-  UnorderedList,
-  Heading,
-  ListItem,
-  Accordion,
-  AccordionItem,
-  AccordionButton,
-  AccordionIcon,
-  AccordionPanel,
-  Spacer,
-} from '@chakra-ui/react'
-import { setSelectedClient } from '../reducers/clientList'
+import { useSelector } from 'react-redux'
+import { Flex, Spacer } from '@chakra-ui/react'
 import { InvoiceCsv } from './invoices/InvoiceCsv'
 import { NewClient } from './clients/NewClient'
+import { ClientSelector } from './clients/ClientSelector'
 
 export function Navbar() {
-  const dispatch = useDispatch()
   const clients = useSelector((state) => state.clientList)
-
-  const orderedCompanyNames = clients.data
-    .map((client) => client.business_name)
-    .sort()
-
-  function handleClick(company) {
-    const myClient = clients.data.find(
-      (client) => client.business_name == company
-    )
-    return dispatch(setSelectedClient(myClient))
-  }
 
   if (clients.loading) {
     return <>Loading ...</>
@@ -36,35 +14,12 @@ export function Navbar() {
 
   return (
     <>
-      <NewClient />
-      <Accordion allowToggle>
-        <AccordionItem>
-          <h2>
-            <AccordionButton>
-              <Heading as="h2" fontSize="2xl" color="#0CA789">
-                Clients
-              </Heading>
-              <AccordionIcon />
-            </AccordionButton>
-          </h2>
-          <AccordionPanel pb={4}>
-            <UnorderedList fontWeight="bold" color="#0CA789" cursor="pointer">
-              {orderedCompanyNames.map((company, i) => (
-                <ListItem
-                  onClick={() => handleClick(company)}
-                  key={i}
-                  listStyleType="none"
-                  boxShadow="sm"
-                >
-                  {company}
-                </ListItem>
-              ))}
-            </UnorderedList>
-          </AccordionPanel>
-        </AccordionItem>
-      </Accordion>
-      <Spacer />
-      <InvoiceCsv />
+      <Flex height={'100%'} direction="column">
+        <ClientSelector />
+        <Spacer />
+        <NewClient />
+        <InvoiceCsv />
+      </Flex>
     </>
   )
 }
